backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
langchain-pinecone
pinecone
pypdf
tiktoken
python-multipart